atteched excel and pdf option and change given in migration and controller
atteched excel and pdf option and change given in migration and controller

+++ app/Http/Controllers/CategoryController.php
@@ -2,8 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\CategoryExport;
 use App\Models\Category;
+use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Http\Request;
+use Maatwebsite\Excel\Facades\Excel;
 use Yajra\DataTables\Facades\DataTables;
 
 class CategoryController extends Controller
@@ -45,7 +48,25 @@ class CategoryController extends Controller
                     }
                 } catch (\Throwable $th) {
                     session()->flash("error", $th->getMessage());
-                        return redirect()->back();
+                    return redirect()->back();
+                }
+            }
+
+            if ($request->edit_status) {
+                try {
+                    $validation = $request->validate([
+                        'status' => "required"
+                    ]);
+                    if ($validation) {
+                        Category::where('id', $request->id)->update([
+                            'status' => $request->status,
+                        ]);
+                        session()->flash("success", "Status Updated Successfully");
+                        return redirect()->route("categories");
+                    }
+                } catch (\Throwable $th) {
+                    session()->flash("error", $th->getMessage());
+                    return redirect()->back();
                 }
             }
         }
@@ -58,13 +79,19 @@ class CategoryController extends Controller
                 return response()->json($category);
             }
 
+            if ($request->get_status) {
+                $id = $request->id;
+                $category = Category::where("id", $id)->first();
+                return response()->json($category);
+            }
+
             if ($request->delete_cate) {
                 $id = $request->id;
                 $category = Category::where("id", $id)->delete();
                 return response()->json($category);
             }
 
-            $data = Category::select(['id', 'name', 'description'])->get();
+            $data = Category::select(['id', 'name', 'description', 'status'])->get();
             return DataTables::of($data)
                 ->addIndexColumn()
                 ->addColumn('action', function ($row) {
@@ -76,6 +103,9 @@ class CategoryController extends Controller
                     <ul class="dropdown-menu">
                         <li>
                             <a href="javascript:void(0)"  class="editRow dropdown-item" data-id="' . $row->id . '">Edit</a>
+                        </li>
+                        <li>
+                            <a href="javascript:void(0)"  class="editStatusRow dropdown-item" data-id="' . $row->id . '">Status</a>
                         </li>
                         <li>
                             <a href="javascript:void(0)" class="deleteRow dropdown-item text-danger" data-id="' . $row->id . '">Delete</a>
@@ -90,5 +120,21 @@ class CategoryController extends Controller
 
 
         return view("category.category");
+    }
+
+    public function CategoryExport(Request $request)
+    {
+        $type = $request->type;
+
+        if ($type == 'excel') {
+            return Excel::download(new CategoryExport, 'categories.xlsx');
+        }
+        if ($type == 'pdf') {
+            $categories = Category::get();
+            $pdf = Pdf::loadView('category.category_pdf', ['categories' => $categories]);
+            return $pdf->download('categories.pdf');
+        }
+
+        // return back()->with('error', 'Invalid export type');
     }
 }
